refactor(executors): move SCSS workspace importer to build-styles executor
Resolves circular dependency between config-utilities and executors packages.

- Move scss-workspace-importer from config-utilities to executors/build-styles
- Update build-styles executor to import from local file
- Remove scss-workspace-importer exports from config-utilities package.json and index.ts
- SCSS importer now colocated with the executor that uses it
